+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Internal
 - Improved Hassette events by bringing event creation into classes, adding helper methods
+  - Separated connected and disconnected events for websocket
 
 ## [0.10.0] - 2025-09-27